Remove extra padding underneath newest message
This was causing there to be an extra blank space between the newest 
message and the chat box, which would show if you scrolled all the way 
down in the messages.

The CSS selector is so specific because we need to override 
`padding-bottom: 40px !important;` which comes from Vue I think. More 
about `!important` 
https://stackoverflow.com/questions/11178673/how-to-override-important